--- redis-pro/Common/Helps.swift
@@ -1,14 +1,0 @@
-//
-//  Helps.swift
-//  redis-pro
-//
-//  Created by chengpanwang on 2021/4/19.
-//
-
-import Foundation
-
-struct Helps {
-    static let PAGE_KEYS = "redis dbsize 命令返回的数量."
-    
-    
-}
