+++ redis-pro/Views/Components/SearchBar.swift
@@ -23,8 +23,6 @@ struct SearchBar: View {
         
         HStack {
             // Search text field
-//            MTextField(value: $keywords, placeholder: placeholder, suffix: "magnifyingglass", onCommit: doAction)
-//                .help(Helps.SEARCH_PATTERN)
             NSearchField(value: $keywords, placeholder: placeholder, onCommit: doAction)
                 .help("HELP_SEARCH_BAR")
         }
