--- redisProTests/CheckUpdateTests.swift
@@ -1,44 +1,0 @@
-//
-//  CheckUpdateTests.swift
-//  redisProTests
-//
-//  Created by chengpanwang on 2021/6/9.
-//
-
-import XCTest
-
-class CheckUpdateTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        
-        if let url = URL(string: "https://gitee.com/chengpan168_admin/redis-pro/raw/dev/.version") {
-            do {
-                let contents = try String(contentsOf: url)
-                print(contents)
-            } catch {
-                // contents could not be loaded
-            }
-        } else {
-            // the URL was bad!
-        }
-        
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
