--- redisProTests/NumberTests.swift
@@ -1,39 +1,0 @@
-//
-//  NumberTests.swift
-//  redisProTests
-//
-//  Created by chengpanwang on 2021/7/21.
-//
-
-import XCTest
-
-class NumberTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-//        for index in (1..<3){
-//            print(index)
-//        }
-        
-        let dict:[String:Any] = ["a" : 1]
-        print(dict["b"] ?? "ssss")
-    }
-    
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
